Styled Login and Create-Account pages - Congregated styling for Login and Create Account pages to make a single Forms.css - Edited JSX to add classes, ids, and made alterations to make styling easier/possible - Created 'Forms' directory and moved Login and Create-Account directories into it
