+++ src/App.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import './App.css';
 import GamePage from './components/Game/GamePage.js';
-import Login from './components/Login/Login.js';
 import useToken from './useToken';
-import CreateAccount from './components/Create-account/CreateAccount.js';
+import Login from './components/Forms/Login/Login.js';
+import CreateAccount from './components/Forms/Create-account/CreateAccount.js';
 
 function App() {
   const { token, setToken } = useToken();
