--- src/components/Create-account/createAccount.css
@@ -1,109 +1,0 @@
-.form {
-    width: 100%;
-    height:100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-
-.form-wrapper {
-    width: 400 px;
-    display: flex;
-    flex-direction: column;
-    padding: 20px 40px;
-    border-radius: 10px;
-    box-shadow: #ffffff;
-}
-
-form{
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-}
-
-h1 {
-    text-align: center;
-    width:100%;
-    color: black;
-    font-weight: lighter;
-}
-
-label {
-    font-size: 0.8em;
-    margin-bottom: 0.25em;
-    color: black;
-    font-weight: lighter;
-}
-
-input {
-    padding: 10px 10px;
-    border-radius: 5px;
-    outline: none;
-    border: 1px solid #cfcfcf
-}
-
-input::placeholder {
-    font-size: 1.2em;
-    font-weight: lighter;
-    color: #999
-}
-
-.username {
-    margin-right: 1%;
-}
-
-.password {
-    margin-left: 1%;
-}
-
-.username,
-.password,
-.email {
-    display:flex;
-    flex-direction: column;
-    margin-bottom: 15px;
-}
-
-.username,
-.password {
-    width:50%;
-}
-
-.email,
-.username,
-.password {
-    width: 100%;
-}
-
-.create-account {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.create-account button {
-    background-color: yellow;
-    color: white;
-    border: 2px solid whitesmoke;
-    width: 100%;
-    margin-top: 1em;
-    padding: 8px 0px;
-    font-size: 1em;
-    font-weight: lighter;
-    letter-spacing:1px;
-    margin-bottom: 0.25em;
-}
-
-.create-account button:hover {
-    color: rgb(255, 0, 13);
-    background-color: rgb(196, 16, 16);
-    border: 2px solid whitesmoke;
-}
-
-.create-account small {
-    color:#999;
-    font-weight: lighter;
-}
